Refactor sim/component.rb to use DSL modules
Extract DSL functionality from sim/component.rb and sim/sequential_component.rb
into separate, reusable modules:

New DSL modules:
- dsl/ports.rb: input, output, wire, parameter DSL
- dsl/structure.rb: instance, port connection DSL
- dsl/vec.rb: vec, input_vec, output_vec DSL
- dsl/bundle.rb: input_bundle, output_bundle DSL
- dsl/codegen.rb: to_ir, to_verilog, to_circt generation
- dsl/sequential_codegen.rb: sequential-specific IR generation

Renames:
- dsl/memory_dsl.rb -> dsl/memory.rb (module: MemoryDSL -> Memory)
- dsl/state_machine_dsl.rb -> dsl/state_machine.rb (module: StateMachineDSL -> StateMachine)

This properly separates concerns:
- DSL: how to define hardware (ports, structure, etc.)
- Simulation: how to execute hardware (propagate, wiring)
- Codegen: how to export hardware (Verilog, FIRRTL)

sim/component.rb now includes DSL modules rather than defining them inline,
making it much smaller and focused on simulation logic.

All 2091 tests pass.
