+++ examples/mos6502/hdl/memory.rb
@@ -1,16 +1,16 @@
 # MOS 6502 Memory Interface - Synthesizable DSL Version
-# Uses MemoryDSL for Verilog export
+# Uses Memory for Verilog export
 # 64KB addressable memory with RAM and ROM regions
 
 require_relative '../../../lib/rhdl'
-require_relative '../../../lib/rhdl/dsl/memory_dsl'
+require_relative '../../../lib/rhdl/dsl/memory'
 
 module MOS6502
   # Synthesizable 64KB memory for 6502
   # RAM: 0x0000 - 0x7FFF (32KB)
   # ROM: 0x8000 - 0xFFFF (32KB)
   class Memory < RHDL::HDL::Component
-    include RHDL::DSL::MemoryDSL
+    include RHDL::DSL::Memory
     include RHDL::DSL::Behavior
 
     # Memory map constants
